Splitting renderer_process into separate modules
